refactor(api): split chat session boundary (#413)
