+++ apps/api/src/chat/__tests__/chat.service.test.ts
@@ -20,6 +20,7 @@ import { validateSync } from 'class-validator';
 import { describe, expect, it, vi } from 'vitest';
 
 import type { AuditEntry, AuditService } from '../../audit/audit.service.js';
+import type { AgentService } from '../../agent/agent.service.js';
 import type { GraphService } from '../../graph/graph.service.js';
 import {
   TEST_GROUP_ID,
@@ -32,6 +33,7 @@ import {
 } from '../../users/__tests__/user-group-service-test-utils.js';
 import { ChatController } from '../chat.controller.js';
 import { ChatService, type ChatStreamEvent } from '../chat.service.js';
+import { ChatSessionBoundaryService } from '../chat-session-boundary.service.js';
 import { UpdateSessionSpacesDto } from '../dto/chat.dto.js';
 
 type ChatSessionRow = typeof chatSessions.$inferSelect;
@@ -155,6 +157,19 @@ describe('ChatService multi-space session', () => {
     ]);
   });
 
+  it('createSession deduplicates duplicate spaceIds while preserving order', async () => {
+    const { service, db } = createServiceContext();
+    db.queueInsert([createSessionRow({ id: 'session-created', space_id: 'space-a' })]);
+
+    await service.createSession(TEST_TENANT_ID, 'space-a', TEST_USER_ID, ['space-a', 'space-a', 'space-b']);
+
+    expect(db.inserts[1]?.table).toBe(chatSessionSpaces);
+    expect(db.inserts[1]?.value).toEqual([
+      expect.objectContaining({ session_id: 'session-created', space_id: 'space-a', position: 0 }),
+      expect.objectContaining({ session_id: 'session-created', space_id: 'space-b', position: 1 }),
+    ]);
+  });
+
   it('updateSessionSpaces replaces membership rows and returns space details', async () => {
     const { service, db } = createServiceContext();
     db.queueSelect([createSessionRow({ id: 'session-1', space_id: 'space-a' })]);
@@ -226,6 +241,10 @@ describe('ChatService multi-space session', () => {
 
     expect(err.getStatus()).toBe(400);
     expect(getHttpExceptionCode(err)).toBe(ErrorCode.VALIDATION_ERROR);
+    expect(getHttpExceptionResponse(err)).toEqual({
+      code: ErrorCode.VALIDATION_ERROR,
+      message: 'space_id must be included in space_ids',
+    });
     expect(db.deletes).toHaveLength(0);
   });
 
@@ -238,6 +257,10 @@ describe('ChatService multi-space session', () => {
 
     expect(err.getStatus()).toBe(400);
     expect(getHttpExceptionCode(err)).toBe(ErrorCode.VALIDATION_ERROR);
+    expect(getHttpExceptionResponse(err)).toEqual({
+      code: ErrorCode.VALIDATION_ERROR,
+      message: 'space_ids must not be empty when provided',
+    });
     expect(db.deletes).toHaveLength(0);
   });
 
@@ -280,6 +303,10 @@ describe('Space scope normalization', () => {
     } catch (err) {
       expect((err as HttpException).getStatus()).toBe(400);
       expect(getHttpExceptionCode(err)).toBe(ErrorCode.VALIDATION_ERROR);
+      expect(getHttpExceptionResponse(err)).toEqual({
+        code: ErrorCode.VALIDATION_ERROR,
+        message: 'At least one Space is required',
+      });
     }
   });
 
@@ -291,6 +318,10 @@ describe('Space scope normalization', () => {
     } catch (err) {
       expect((err as HttpException).getStatus()).toBe(400);
       expect(getHttpExceptionCode(err)).toBe(ErrorCode.VALIDATION_ERROR);
+      expect(getHttpExceptionResponse(err)).toEqual({
+        code: ErrorCode.VALIDATION_ERROR,
+        message: 'space_ids must not be empty when provided',
+      });
       return;
     }
 
@@ -305,6 +336,10 @@ describe('Space scope normalization', () => {
     } catch (err) {
       expect((err as HttpException).getStatus()).toBe(400);
       expect(getHttpExceptionCode(err)).toBe(ErrorCode.VALIDATION_ERROR);
+      expect(getHttpExceptionResponse(err)).toEqual({
+        code: ErrorCode.VALIDATION_ERROR,
+        message: 'space_id must be included in space_ids',
+      });
       return;
     }
 
@@ -313,6 +348,111 @@ describe('Space scope normalization', () => {
 });
 
 describe('ChatService multi-space sessions', () => {
+  it('streamCompletion creates duplicate requested scope once after permission succeeds', async () => {
+    const { service, db } = createServiceContext();
+    db.queueSelect([createSpaceRow({ id: 'space-a', name: 'Space A' })]);
+    db.queueSelect([createSpaceRow({ id: 'space-b', name: 'Space B' })]);
+    db.queueSelect([createModelRow({ model_type: 'chat' })]);
+    db.queueInsert([createSessionRow({ id: 'session-created', space_id: 'space-a' })]);
+    db.queueSelect([createSessionRow({ id: 'session-created', space_id: 'space-a' })]);
+    db.queueSelect([]);
+    db.queueInsert([createMessageRow({ id: 'user-message', role: 'user' })]);
+    db.queueSelect([]);
+    db.queueSelect([]);
+    db.queueInsert([createMessageRow({ id: 'assistant-no-hit', role: 'assistant', content: NO_HIT_MESSAGE })]);
+
+    const events = await collectEvents(
+      await service.streamCompletion({
+        tenantId: TEST_TENANT_ID,
+        spaceId: 'space-a',
+        spaceIds: ['space-a', 'space-a', 'space-b'],
+        userId: TEST_USER_ID,
+        userGroupIds: [],
+        actorRole: 'viewer',
+        spacePermissions: { 'space-a': ['chat:use'], 'space-b': ['chat:use'] },
+        message: 'hello',
+      }),
+    );
+
+    expect(events[0]).toEqual({ type: 'session', session_id: 'session-created' });
+    expect(db.inserts.find((insert) => insert.table === chatSessionSpaces)?.value).toEqual([
+      expect.objectContaining({ session_id: 'session-created', space_id: 'space-a', position: 0 }),
+      expect.objectContaining({ session_id: 'session-created', space_id: 'space-b', position: 1 }),
+    ]);
+  });
+
+  it.each([
+    {
+      name: 'explicit empty scope',
+      input: { spaceId: 'space-a', spaceIds: [] },
+      message: 'space_ids must not be empty when provided',
+    },
+    {
+      name: 'missing implicit scope',
+      input: {},
+      message: 'At least one Space is required',
+    },
+    {
+      name: 'over-limit scope',
+      input: { spaceIds: Array.from({ length: 11 }, (_, index) => `space-${index + 1}`) },
+      message: 'At most 10 Spaces can be selected',
+    },
+    {
+      name: 'mismatched primary scope',
+      input: { spaceId: 'space-c', spaceIds: ['space-a', 'space-b'] },
+      message: 'space_id must be included in space_ids',
+    },
+  ])('rejects $name before any session, membership, or message write', async ({ input, message }) => {
+    const { service, db } = createServiceContext();
+
+    const err = await getRejectedHttpException(
+      service.streamCompletion({
+        tenantId: TEST_TENANT_ID,
+        userId: TEST_USER_ID,
+        userGroupIds: [],
+        message: 'hello',
+        ...input,
+      }),
+    );
+
+    expect(err.getStatus()).toBe(400);
+    expect(getHttpExceptionResponse(err)).toEqual({
+      code: ErrorCode.VALIDATION_ERROR,
+      message,
+    });
+    expect(db.inserts).toHaveLength(0);
+    expect(db.deletes).toHaveLength(0);
+    expect(db.updates).toHaveLength(0);
+  });
+
+  it('rejects unauthorized requested spaces before creating a session or messages', async () => {
+    const { service, db } = createServiceContext();
+    db.queueSelect([createSpaceRow({ id: 'space-a' })]);
+    db.queueSelect([createSpaceRow({ id: 'space-b' })]);
+
+    const err = await getRejectedHttpException(
+      service.streamCompletion({
+        tenantId: TEST_TENANT_ID,
+        spaceId: 'space-a',
+        spaceIds: ['space-a', 'space-b'],
+        userId: TEST_USER_ID,
+        userGroupIds: [],
+        actorRole: 'viewer',
+        spacePermissions: { 'space-a': ['chat:use'], 'space-b': ['space:view'] },
+        message: 'hello',
+      }),
+    );
+
+    expect(err.getStatus()).toBe(403);
+    expect(getHttpExceptionResponse(err)).toEqual({
+      code: ErrorCode.PERMISSION_DENIED,
+      message: 'Permission denied',
+    });
+    expect(db.inserts).toHaveLength(0);
+    expect(db.deletes).toHaveLength(0);
+    expect(db.updates).toHaveLength(0);
+  });
+
   it('session continuation with same scope succeeds', async () => {
     const { service, db } = createServiceContext();
     queueExistingCompletion(db, ['space-a', 'space-b']);
@@ -354,6 +494,8 @@ describe('ChatService multi-space sessions', () => {
         (insert) => insert.table === chatMessages && isRecordForTest(insert.value) && insert.value.role === 'user',
       ),
     ).toBe(true);
+    expect(db.inserts.some((insert) => insert.table === chatSessionSpaces)).toBe(false);
+    expect(db.deletes.some((deletion) => deletion.table === chatSessionSpaces)).toBe(false);
   });
 
   it('session continuation with only space_id uses stored membership', async () => {
@@ -447,6 +589,10 @@ describe('ChatService multi-space sessions', () => {
 
     expect(err.getStatus()).toBe(404);
     expect(getHttpExceptionCode(err)).toBe(ErrorCode.SESSION_NOT_FOUND);
+    expect(getHttpExceptionResponse(err)).toEqual({
+      code: ErrorCode.SESSION_NOT_FOUND,
+      message: 'Chat session was not found',
+    });
   });
 
   it('deleteSession from member space succeeds with cascade', async () => {
@@ -460,6 +606,23 @@ describe('ChatService multi-space sessions', () => {
     await expect(service.deleteSession(TEST_TENANT_ID, 'session-1', TEST_USER_ID, 'space-b')).resolves.toEqual({
       deleted: true,
     });
+    expect(db.deletes[0]?.table).toBe(chatSessions);
+  });
+
+  it('deleteSession returns deleted even when Agent close fails', async () => {
+    const close = vi.fn(() => Promise.reject(new Error('close failed')));
+    const agentService = {
+      close,
+    } as unknown as AgentService;
+    const { service, db } = createServiceContext({ agentService });
+    db.queueSelect([createSessionRow({ space_id: 'space-a' })]);
+    db.queueSelect([createSessionSpaceRow({ space_id: 'space-a', position: 0 })]);
+
+    await expect(service.deleteSession(TEST_TENANT_ID, 'session-1', TEST_USER_ID, 'space-a')).resolves.toEqual({
+      deleted: true,
+    });
+
+    expect(close).toHaveBeenCalledWith('session-1');
     expect(db.deletes[0]?.table).toBe(chatSessions);
   });
 
@@ -490,6 +653,10 @@ describe('ChatService multi-space sessions', () => {
 
     expect(err.getStatus()).toBe(403);
     expect(getHttpExceptionCode(err)).toBe(ErrorCode.PERMISSION_DENIED);
+    expect(getHttpExceptionResponse(err)).toEqual({
+      code: ErrorCode.PERMISSION_DENIED,
+      message: 'Permission denied',
+    });
     expect(db.inserts).toHaveLength(0);
   });
 
@@ -513,6 +680,75 @@ describe('ChatService multi-space sessions', () => {
       message: 'Space not found',
     });
     expect(db.inserts).toHaveLength(0);
+  });
+
+  it('returns SESSION_NOT_FOUND for nonexistent session entrypoints without writes', async () => {
+    const getContext = createServiceContext();
+    getContext.db.queueSelect([]);
+    const getErr = await getRejectedHttpException(
+      getContext.service.getSession(TEST_TENANT_ID, 'missing-session', TEST_USER_ID, 'space-a'),
+    );
+    expect(getErr.getStatus()).toBe(404);
+    expect(getHttpExceptionResponse(getErr)).toEqual({
+      code: ErrorCode.SESSION_NOT_FOUND,
+      message: 'Chat session was not found',
+    });
+    expect(getContext.db.inserts).toHaveLength(0);
+
+    const updateContext = createServiceContext();
+    updateContext.db.queueSelect([]);
+    const updateErr = await getRejectedHttpException(
+      updateContext.service.updateSessionSpaces(
+        TEST_TENANT_ID,
+        'missing-session',
+        TEST_USER_ID,
+        'space-a',
+        ['space-a'],
+      ),
+    );
+    expect(updateErr.getStatus()).toBe(404);
+    expect(getHttpExceptionResponse(updateErr)).toEqual({
+      code: ErrorCode.SESSION_NOT_FOUND,
+      message: 'Chat session was not found',
+    });
+    expect(updateContext.db.inserts).toHaveLength(0);
+    expect(updateContext.db.deletes).toHaveLength(0);
+    expect(updateContext.db.updates).toHaveLength(0);
+
+    const deleteContext = createServiceContext();
+    deleteContext.db.queueSelect([]);
+    const deleteErr = await getRejectedHttpException(
+      deleteContext.service.deleteSession(TEST_TENANT_ID, 'missing-session', TEST_USER_ID, 'space-a'),
+    );
+    expect(deleteErr.getStatus()).toBe(404);
+    expect(getHttpExceptionResponse(deleteErr)).toEqual({
+      code: ErrorCode.SESSION_NOT_FOUND,
+      message: 'Chat session was not found',
+    });
+    expect(deleteContext.db.inserts).toHaveLength(0);
+    expect(deleteContext.db.deletes).toHaveLength(0);
+
+    const streamContext = createServiceContext();
+    streamContext.db.queueSelect([createModelRow({ model_type: 'chat' })]);
+    streamContext.db.queueSelect([]);
+    const streamErr = await getRejectedHttpException(
+      streamContext.service.streamCompletion({
+        tenantId: TEST_TENANT_ID,
+        spaceId: 'space-a',
+        sessionId: 'missing-session',
+        userId: TEST_USER_ID,
+        userGroupIds: [],
+        message: 'hello',
+      }),
+    );
+    expect(streamErr.getStatus()).toBe(404);
+    expect(getHttpExceptionResponse(streamErr)).toEqual({
+      code: ErrorCode.SESSION_NOT_FOUND,
+      message: 'Chat session was not found',
+    });
+    expect(streamContext.db.inserts).toHaveLength(0);
+    expect(streamContext.db.deletes).toHaveLength(0);
+    expect(streamContext.db.updates).toHaveLength(0);
   });
 });
 
@@ -1399,6 +1635,7 @@ type ServiceContextOptions = {
   chatProvider?: ScriptedChatProvider;
   embeddingProvider?: ScriptedEmbeddingProvider;
   graphService?: GraphService;
+  agentService?: AgentService;
 };
 
 function createServiceContext(options: ServiceContextOptions = {}): {
@@ -1407,6 +1644,7 @@ function createServiceContext(options: ServiceContextOptions = {}): {
   audit: { push: ReturnType<typeof vi.fn<(entry: AuditEntry) => void>> };
   chatFactory: ReturnType<typeof vi.fn>;
   embeddingFactory: ReturnType<typeof vi.fn>;
+  sessionBoundary: ChatSessionBoundaryService;
 } {
   const db = new ScriptedChatDb();
   const audit = {
@@ -1416,16 +1654,19 @@ function createServiceContext(options: ServiceContextOptions = {}): {
   const embeddingProvider = options.embeddingProvider ?? new ScriptedEmbeddingProvider();
   const chatFactory = vi.fn(() => chatProvider);
   const embeddingFactory = vi.fn(() => embeddingProvider);
+  const sessionBoundary = new ChatSessionBoundaryService(db.asDrizzle());
   const service = new ChatService(
     db.asDrizzle(),
     audit as unknown as AuditService,
     chatFactory,
     embeddingFactory,
-    undefined,
+    options.agentService,
     options.graphService,
+    undefined,
+    sessionBoundary,
   );
 
-  return { service, db, audit, chatFactory, embeddingFactory };
+  return { service, db, audit, chatFactory, embeddingFactory, sessionBoundary };
 }
 
 function queuePreparedCompletion(db: ScriptedChatDb, options: { space?: SpaceRow } = {}): void {
@@ -1864,7 +2105,7 @@ function normalizeSpaceScopeForTest(
   service: ChatService,
   input: { space_id?: string; space_ids?: string[] },
 ): string[] {
-  return (service as unknown as { normalizeSpaceScope: (dto: typeof input) => string[] }).normalizeSpaceScope(input);
+  return (service as unknown as { sessionBoundary: ChatSessionBoundaryService }).sessionBoundary.normalizeSpaceScope(input);
 }
 
 function isRecordForTest(value: unknown): value is Record<string, unknown> {
