+++ apps/api/src/chat/chat.module.ts
@@ -7,12 +7,14 @@ import { DrizzleModule } from '../database/drizzle.module.js';
 import { GraphModule } from '../graph/graph.module.js';
 import { ModelConfigModule } from '../models/model-config.module.js';
 import { CHAT_PROVIDER_FACTORY, EMBEDDING_PROVIDER_FACTORY, ChatService } from './chat.service.js';
+import { ChatSessionBoundaryService } from './chat-session-boundary.service.js';
 import { ChatController } from './chat.controller.js';
 
 @Module({
   imports: [DrizzleModule, AuditModule, AgentModule, GraphModule, ModelConfigModule],
   controllers: [ChatController],
   providers: [
+    ChatSessionBoundaryService,
     ChatService,
     {
       provide: CHAT_PROVIDER_FACTORY,
